Оптимизация кода

+++ helpers/searchhelpers.py
@@ -11,7 +11,7 @@ def build_hotels_list(props_list: list) -> list:
         {
         "id": get_value(i_prop, "id"),
         "name": get_value(i_prop, "name"),
-        "score": float(get_value(i_prop, "score")),
+        "score": int(get_value(i_prop, "score")),
         "amount": round(float(get_value(i_prop, "amount")), 2)
     }
     for i_prop in props_list
